+++ config.php
@@ -7,6 +7,8 @@ use humhub\modules\external_calendar\Module;
 use humhub\commands\IntegrityController;
 use humhub\modules\calendar\interfaces\CalendarService;
 use humhub\modules\content\widgets\WallEntryLinks;
+use humhub\modules\stream\models\WallStreamQuery;
+use humhub\modules\stream\widgets\WallStreamFilterNavigation;
 
 return [
     'id' => 'external_calendar',
@@ -20,6 +22,8 @@ return [
         ['class' => CronController::class, 'event' => CronController::EVENT_ON_DAILY_RUN, 'callback' => [Events::class, 'onCronDailyRun']],
         ['class' => IntegrityController::class, 'event' => IntegrityController::EVENT_ON_RUN, 'callback' => [Events::class, 'onIntegrityCheck']],
         ['class' => WallEntryLinks::class, 'event' => 'init', 'callback' => [Events::class, 'onWallEntryLinks']],
+        ['class' => WallStreamQuery::class, 'event' =>  WallStreamQuery::EVENT_BEFORE_FILTER, 'callback' => [Events::class, 'onStreamFilterBeforeFilter']],
+        ['class' => WallStreamFilterNavigation::class, 'event' =>  WallStreamFilterNavigation::EVENT_BEFORE_RUN, 'callback' => [Events::class, 'onStreamFilterBeforeRun']],
     ],
 ];
 ?>